+++ src/main/java/domain/piece/Piece.java
@@ -1,5 +1,7 @@
 package domain.piece;
 
+import domain.piece.move.Coordinate;
+
 public interface Piece {
 
     boolean isReachableByRule(final Coordinate startCoordinate, final Coordinate endCoordinate);
