+++ src/main/java/domain/piece/pawn/WhitePawn.java
@@ -1,16 +1,21 @@
-package domain.piece;
+package domain.piece.pawn;
+
+import domain.piece.move.Coordinate;
+import domain.piece.move.Direction;
+import domain.piece.move.Inclination;
 
 import java.util.List;
 
-public final class BlackPawn extends Pawn {
+public final class WhitePawn extends Pawn {
 
     private static final Direction DIRECTION = new Direction(List.of(
-            Inclination.NEGATIVE_INFINITY
+            Inclination.POSITIVE_INFINITY
     ));
 
     @Override
     public boolean isReachableByRuleWhenFirstMove(final Coordinate startCoordinate, final Coordinate endCoordinate) {
         Inclination inclination = Inclination.of(startCoordinate.getInclination(endCoordinate));
+
         return DIRECTION.canBeDirectionOf(inclination) &&
                 startCoordinate.hasDistanceLessThanTwo(endCoordinate);
     }
