+++ src/main/java/domain/piece/sliding/Bishop.java
@@ -1,0 +1,22 @@
+package domain.piece.sliding;
+
+import domain.piece.Piece;
+import domain.piece.move.Coordinate;
+import domain.piece.move.Direction;
+import domain.piece.move.Inclination;
+
+import java.util.List;
+
+public final class Bishop implements Piece {
+
+    private static final Direction DIRECTION = new Direction(List.of(
+            Inclination.ONE, Inclination.MINUS_ONE
+    ));
+
+    @Override
+    public boolean isReachableByRule(final Coordinate startCoordinate, final Coordinate endCoordinate) {
+        Inclination inclination = Inclination.of(startCoordinate.getInclination(endCoordinate));
+
+        return DIRECTION.canBeDirectionOf(inclination);
+    }
+}
