+++ src/test/java/controller/CoordinateAdapterTest.java
@@ -1,6 +1,6 @@
 package controller;
 
-import domain.piece.Coordinate;
+import domain.piece.move.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
