+++ src/test/java/domain/board/BoardTest.java
@@ -1,6 +1,6 @@
 package domain.board;
 
-import domain.piece.Coordinate;
+import domain.piece.move.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
