+++ src/test/java/domain/piece/EmptyPieceTest.java
@@ -1,5 +1,6 @@
 package domain.piece;
 
+import domain.piece.move.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
