+++ src/test/java/domain/piece/move/InclinationTest.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain.piece.move;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
